seperated the memberhip code in components
